% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_genome.R
\name{download_genome}
\alias{download_genome}
\title{Download a genome from RefSeq}
\usage{
download_genome(species, index_dir)
}
\arguments{
\item{species}{Name of the species to download}

\item{index_dir}{Directory in which downloaded genomes should be cached}
}
\description{
Download a genome from RefSeq
}
