% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqdecontamr-package.R
\docType{package}
\name{fqdecontamr-package}
\alias{fqdecontamr}
\alias{fqdecontamr-package}
\title{fqdecontamr: Remove contaminant reads from fastq files}
\description{
Given a list of contaminant taxa identified by the decontam
  software, and taxonomic annotations provided by MetaPhlAn2, this package can
  remove the reads belonging to contaminant taxa from a fastq or set of fastqs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wilkox/fqdecontamr}
  \item Report bugs at \url{https://github.com/wilkox/fqdecontamr/issues}
}

}
\author{
\strong{Maintainer}: David Wilkins \email{david@wilkox.org}

}
\keyword{internal}
