% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_bowtie_index.R
\name{prepare_bowtie_index}
\alias{prepare_bowtie_index}
\title{Prepare a bowtie2 index}
\usage{
prepare_bowtie_index(fasta_path, dir, species, ...)
}
\arguments{
\item{fasta_path}{Path of the fasta file from which to prepare the index}

\item{dir}{Directory in which to store the index}

\item{species}{Name of the contaminant species}

\item{...}{Other arguments that will be ignored}
}
\description{
Prepare a bowtie2 index
}
