% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_genomes.R
\name{prepare_bowtie_indices}
\alias{prepare_bowtie_indices}
\title{Prepare bowtie2 indices for aligment}
\usage{
prepare_bowtie_indices(contaminants, index_dir)
}
\arguments{
\item{contaminants}{[Tibble][tibble::tibble-package] listing contaminants
identified by decontam, with at least a 'species' column}

\item{index_dir}{Directory in which to store the downloaded genomes and bowtie2
indices}
}
\description{
Given a list of contaminant species, attempt to download genomes for each
contaminant and prepare a bowtie2 index for alignment
}
